Refactor logcap tests
